Move TrustedUpdater to optimizers.trusted_updater.py
Add truested_updater_test.py
Address comments
